fix #605 Chess.com v3: offline mode for TT cause blinking issues
fix Stats switching logic via spinner
fix radioButton in GraphView

fix lesson_complete_percentage issue

+++ src/com/chess/backend/entity/api/stats/UserLessonsStatsData.java
@@ -78,7 +78,7 @@ public class UserLessonsStatsData {
 	public class Stats {
 		private int lessons_tried;
 		private int total_lesson_count;
-		private int lesson_complete_percentage;
+		private float lesson_complete_percentage;
 		private long total_training_seconds;
 		private Score score;
 
@@ -90,7 +90,7 @@ public class UserLessonsStatsData {
 			return total_lesson_count;
 		}
 
-		public int getLessonCompletePercentage() {
+		public float getLessonCompletePercentage() {
 			return lesson_complete_percentage;
 		}
